[charts] Extract bar and line plot logic into reusable hooks
