+++ packages/x-charts/src/LineChart/AreaPlot.tsx
@@ -2,7 +2,6 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
 import { styled } from '@mui/material/styles';
-import { area as d3Area } from '@mui/x-charts-vendor/d3-shape';
 import {
   AreaElement,
   areaElementClasses,
@@ -10,16 +9,11 @@ import {
   AreaElementSlotProps,
   AreaElementSlots,
 } from './AreaElement';
-import { getValueToPositionMapper } from '../hooks/useScale';
-import { getCurveFactory } from '../internals/getCurve';
-import { isBandScale } from '../internals/isBandScale';
-import { DEFAULT_X_AXIS_KEY } from '../constants';
 import { LineItemIdentifier } from '../models/seriesType/line';
-import { useLineSeriesContext } from '../hooks/useLineSeries';
 import { useSkipAnimation } from '../hooks/useSkipAnimation';
-import { useChartGradientIdBuilder } from '../hooks/useChartGradientId';
 import { useXAxes, useYAxes } from '../hooks/useAxis';
 import { useInternalIsZoomInteracting } from '../internals/plugins/featurePlugins/useChartCartesianAxis/useInternalIsZoomInteracting';
+import { useAreaPlotData } from './useAreaPlotData';
 
 export interface AreaPlotSlots extends AreaElementSlots {}
 
@@ -49,138 +43,10 @@ const AreaPlotRoot = styled('g', {
 });
 
 const useAggregatedData = () => {
-  const seriesData = useLineSeriesContext();
-  const { xAxis, xAxisIds } = useXAxes();
-  const { yAxis, yAxisIds } = useYAxes();
-  const getGradientId = useChartGradientIdBuilder();
+  const { xAxis: xAxes } = useXAxes();
+  const { yAxis: yAxes } = useYAxes();
 
-  // This memo prevents odd line chart behavior when hydrating.
-  const allData = React.useMemo(() => {
-    if (seriesData === undefined) {
-      return [];
-    }
-
-    const { series, stackingGroups } = seriesData;
-
-    const defaultXAxisId = xAxisIds[0];
-    const defaultYAxisId = yAxisIds[0];
-
-    return stackingGroups.flatMap(({ ids: groupIds }) => {
-      return [...groupIds]
-        .reverse() // Revert stacked area for a more pleasant animation
-        .map((seriesId) => {
-          const {
-            xAxisId = defaultXAxisId,
-            yAxisId = defaultYAxisId,
-            stackedData,
-            data,
-            connectNulls,
-            baseline,
-            curve,
-            strictStepCurve,
-          } = series[seriesId];
-
-          const xScale = xAxis[xAxisId].scale;
-          const xPosition = getValueToPositionMapper(xScale);
-          const yScale = yAxis[yAxisId].scale;
-          const xData = xAxis[xAxisId].data;
-
-          const gradientId: string | undefined =
-            (yAxis[yAxisId].colorScale && getGradientId(yAxisId)) ||
-            (xAxis[xAxisId].colorScale && getGradientId(xAxisId)) ||
-            undefined;
-
-          if (process.env.NODE_ENV !== 'production') {
-            if (xData === undefined) {
-              throw new Error(
-                `MUI X Charts: ${
-                  xAxisId === DEFAULT_X_AXIS_KEY
-                    ? 'The first `xAxis`'
-                    : `The x-axis with id "${xAxisId}"`
-                } should have data property to be able to display a line plot.`,
-              );
-            }
-            if (xData.length < stackedData.length) {
-              throw new Error(
-                `MUI X Charts: The data length of the x axis (${xData.length} items) is lower than the length of series (${stackedData.length} items).`,
-              );
-            }
-          }
-
-          const shouldExpand = curve?.includes('step') && !strictStepCurve && isBandScale(xScale);
-
-          const formattedData: {
-            x: any;
-            y: [number, number];
-            nullData: boolean;
-            isExtension?: boolean;
-          }[] =
-            xData?.flatMap((x, index) => {
-              const nullData = data[index] == null;
-              if (shouldExpand) {
-                const rep = [{ x, y: stackedData[index], nullData, isExtension: false }];
-                if (!nullData && (index === 0 || data[index - 1] == null)) {
-                  rep.unshift({
-                    x: (xScale(x) ?? 0) - (xScale.step() - xScale.bandwidth()) / 2,
-                    y: stackedData[index],
-                    nullData,
-                    isExtension: true,
-                  });
-                }
-                if (!nullData && (index === data.length - 1 || data[index + 1] == null)) {
-                  rep.push({
-                    x: (xScale(x) ?? 0) + (xScale.step() + xScale.bandwidth()) / 2,
-                    y: stackedData[index],
-                    nullData,
-                    isExtension: true,
-                  });
-                }
-                return rep;
-              }
-              return { x, y: stackedData[index], nullData };
-            }) ?? [];
-
-          const d3Data = connectNulls ? formattedData.filter((d) => !d.nullData) : formattedData;
-
-          const areaPath = d3Area<{
-            x: any;
-            y: [number, number];
-            nullData: boolean;
-            isExtension?: boolean;
-          }>()
-            .x((d) => (d.isExtension ? d.x : xPosition(d.x)))
-            .defined((d) => connectNulls || !d.nullData || !!d.isExtension)
-            .y0((d) => {
-              if (typeof baseline === 'number') {
-                return yScale(baseline)!;
-              }
-              if (baseline === 'max') {
-                return yScale.range()[1];
-              }
-              if (baseline === 'min') {
-                return yScale.range()[0];
-              }
-
-              const value = d.y && yScale(d.y[0])!;
-              if (Number.isNaN(value)) {
-                return yScale.range()[0];
-              }
-              return value;
-            })
-            .y1((d) => d.y && yScale(d.y[1])!);
-
-          const d = areaPath.curve(getCurveFactory(curve))(d3Data) || '';
-          return {
-            ...series[seriesId],
-            gradientId,
-            d,
-            seriesId,
-          };
-        });
-    });
-  }, [seriesData, xAxisIds, yAxisIds, xAxis, yAxis, getGradientId]);
-
-  return allData;
+  return useAreaPlotData(xAxes, yAxes);
 };
 
 /**
