+++ packages/x-charts/src/internals/index.ts
@@ -10,7 +10,10 @@ export { useInteractionItemProps } from '../hooks/useInteractionItemProps';
 export { useDrawingArea } from '../hooks/useDrawingArea';
 export { useScatterChartProps } from '../ScatterChart/useScatterChartProps';
 export { useLineChartProps } from '../LineChart/useLineChartProps';
+export { useAreaPlotData } from '../LineChart/useAreaPlotData';
+export { useLinePlotData } from '../LineChart/useLinePlotData';
 export { useBarChartProps } from '../BarChart/useBarChartProps';
+export { useBarPlotData } from '../BarChart/useBarPlotData';
 export { useRadarChartProps } from '../RadarChart/useRadarChartProps';
 export * from '../ChartContainer/useChartContainerProps';
 export * from '../ChartDataProvider/useChartDataProviderProps';
